fine grained support for enabling mappings
Create a function that determines whether or not the clipboard mappings are
in effect. This includes the original "!has('clipboard')" test. It adds a
test for Vim on Mac OS X built with X11/GTK2 support running in a terminal.
It also adds a global var the user can set to force all key mappings. This
is similar to change proposed by ~tyru a year ago but I chose a different
name for the variable.